+++ index.js
@@ -1,5 +1,1 @@
-const Controller = require('./dist/Controller').default;
-const Utils = require('./dist/misc/utils').default;
-
-exports.Controller = Controller;
-exports.Utils = Utils;
+export const Controller = require('./dist/Controller').default;