+++ src/evm/Transaction.ts
@@ -1,26 +1,23 @@
 import * as JSONBig from 'json-bigint'
 
-import utils from "../misc/utils";
-
-import {TX, TXReceipt} from "../misc/Interfaces";
+import {TX, TXReceipt} from "./utils/Interfaces";
 
 import Controller from "../Controller";
 
 
 export default class Transaction {
-    public tx: TX;
+
     public receipt: TXReceipt;
 
     /**
      * Transaction object to be sent or called.
      *
-     * @param {TX} options The transaction options eg. gas, gas price, value...
-     * @param {boolean} constant If the transaction is constant
-     * @param {Function} unpackfn If constant - unpack function
-     * @param {Controller} controller The controller class
+     * @param {TX} _tx - The transaction options eg. gas, gas price, value...
+     * @param {boolean} constant - If the transaction is constant
+     * @param {Function} unpackfn - If constant - unpack function
+     * @param {Controller} controller - The controller class
      */
-    constructor(options: TX, readonly constant: boolean, readonly unpackfn: Function, readonly controller: Controller) {
-        this.tx = options;
+    constructor(private _tx: TX, readonly constant: boolean, readonly unpackfn: Function, readonly controller: Controller) {
         this.receipt = undefined;
 
         if (!constant)
@@ -32,34 +29,32 @@ export default class Transaction {
      *
      * This function will mutate the state of the EVM.
      *
-     * @param {Object} options
+     * @param {Object} options - The options to pass to the transaction
      */
     send(options?: { to?: string, from?: string, value?: number, gas?: number, gasPrice?: number }): any {
         if (!this.constant) {
             if (options) {
-                this.tx.to = options.to || this.tx.to;
-                this.tx.from = options.from || this.tx.from;
-                this.tx.gas = options.gas || this.tx.gas;
-                this.tx.value = options.value || this.tx.value;
+                this._tx.to = options.to || this._tx.to;
+                this._tx.from = options.from || this._tx.from;
+                this._tx.gas = options.gas || this._tx.gas;
+                this._tx.value = options.value || this._tx.value;
 
                 if (options.gasPrice !== undefined && options.gasPrice >= 0) {
-                    this.tx.gasPrice = options.gasPrice;
+                    this._tx.gasPrice = options.gasPrice;
                 }
             }
 
-            utils.log(utils.fgGreen, JSONBig.stringify(this.tx, null, 2));
-
-            if (this.tx.gas != null && this.tx.gasPrice != null) {
-                return this.controller.api.sendTx(JSONBig.stringify(this.tx))
+            if (this._tx.gas != null && this._tx.gasPrice != null) {
+                return this.controller.api.sendTx(JSONBig.stringify(this._tx))
                     .then((res: string) => {
                         let response: { txHash: string } = JSONBig.parse(res);
                         return response.txHash
                     })
                     .then((txHash) => {
-                        return utils.sleep(2000).then(() => {
-                            utils.log(utils.fgBlue, 'Requesting Receipt');
-                            return this.controller.api.getReceipt(txHash)
-                        })
+                        return new Promise((resolve) => setTimeout(resolve, 2000))
+                            .then(() => {
+                                return this.controller.api.getReceipt(txHash)
+                            })
                     })
                     .then((resp: string) => {
                         this.receipt = JSONBig.parse(resp);
@@ -78,24 +73,23 @@ export default class Transaction {
      *
      * This function will not mutate the state of the EVM.
      *
+     * @param {Object} options - The options to pass to the transaction
      */
     call(options?: { to?: string, from?: string, value?: number, gas?: number, gasPrice?: number }) {
         if (this.constant) {
             if (options) {
-                this.tx.to = options.to || this.tx.to;
-                this.tx.from = options.from || this.tx.from;
-                this.tx.gas = options.gas || this.tx.gas;
-                this.tx.value = options.value || this.tx.value;
+                this._tx.to = options.to || this._tx.to;
+                this._tx.from = options.from || this._tx.from;
+                this._tx.gas = options.gas || this._tx.gas;
+                this._tx.value = options.value || this._tx.value;
 
                 if (options.gasPrice !== undefined && options.gasPrice >= 0) {
-                    this.tx.gasPrice = options.gasPrice;
+                    this._tx.gasPrice = options.gasPrice;
                 }
             }
 
-            utils.log(utils.fgGreen, JSONBig.stringify(this.tx, null, 2));
-
-            if (this.tx.gas != null && this.tx.gasPrice != null) {
-                return this.controller.api.call(JSONBig.stringify(this.tx))
+            if (this._tx.gas != null && this._tx.gasPrice != null) {
+                return this.controller.api.call(JSONBig.stringify(this._tx))
                     .then((response) => {
                         return JSONBig.parse(response);
                     })
@@ -111,68 +105,77 @@ export default class Transaction {
     }
 
     /**
+     * Return transaction as string.
+     *
+     * @returns {string} Transaction as string
+     */
+    toString(): string {
+        return JSONBig.stringify(this._tx);
+    }
+
+    /**
      * Sets the from of the transaction.
      *
-     * @param {string} from The from address
+     * @param {string} from - The from address
      * @returns {Transaction} The transaction
      */
-    from(from: string): Transaction {
-        this.tx.from = from;
+    from(from: string): this {
+        this._tx.from = from;
         return this
     }
 
     /**
      * Sets the to of the transaction.
      *
-     * @param {string} to The to address
+     * @param {string} to - The to address
      * @returns {Transaction} The transaction
      */
-    to(to: string): Transaction {
-        this.tx.to = to;
+    to(to: string): this {
+        this._tx.to = to;
         return this
     }
 
     /**
      * Sets the value of the transaction.
      *
-     * @param {number} value The value of tx
+     * @param {number} value - The value of tx
      * @returns {Transaction} The transaction
      */
-    value(value: number): Transaction {
-        this.tx.value = value;
+    value(value: number): this {
+        this._tx.value = value;
         return this
     }
 
     /**
      * Sets the gas of the transaction.
      *
-     * @param {number} gas The gas of tx
+     * @param {number} gas - The gas of tx
      * @returns {Transaction} The transaction
      */
-    gas(gas: number): Transaction {
-        this.tx.gas = gas;
+    gas(gas: number): this {
+        this._tx.gas = gas;
         return this
     }
 
     /**
      * Sets the gas price of the transaction.
      *
-     * @param {number} gasPrice The gas price of tx
+     * @param {number} gasPrice - The gas price of tx
      * @returns {Transaction} The transaction
      */
-    gasPrice(gasPrice: number): Transaction {
-        this.tx.gasPrice = gasPrice;
+    gasPrice(gasPrice: number): this {
+        this._tx.gasPrice = gasPrice;
         return this
     }
 
     /**
      * Sets the data of the transaction.
      *
-     * @param {number} data The data of tx
+     * @param {number} data - The data of tx
      * @returns {Transaction} The transaction
      */
-    data(data: string): Transaction {
-        this.tx.data = data;
+    data(data: string): this {
+        this._tx.data = data;
         return this
     }
 }