+++ CHANGELOG.md
@@ -2,17 +2,25 @@
 
 FEATURES:
 
+* call() and send() transactions
+* API to handle interactions with transactions
 * Contract abstraction
-* Transfer
-* HTTP Handler
+    * Solidity function abstraction
+* Transfer tokens
+* Promised wrapped HTTP handler
+* Default options for top level class
+* Account handling (WiP)
+* Wallet handling (WiP)
 
 IMPROVEMENTS:
 
 * Testing Capabilities
 * Documentation for classes and functions
 * Functions to set transaction attributes
-* User experience improvements (SolidityContract.ts, SolidityFunction.ts)
-* Type definitions (WiP)
+* User experience improvements SolidityContract.ts
+* User experience improvements SolidityFunction.ts
+* Type definitions
+
 
 SECURITY: